Updating buffer to preload with saved states if available and match size of states.
Previously the buffer was much larger than the number of states collected by "refresh"
due to it expecting each input to be for an LLM containing multiple tokens.
In reality, only one new data point is collected every iteration in refresh,
which is now reflected by the buffer having the same size as the number of states.
Furthermore, when half the buffer is refreshed the state dict will drop samples to remain the same size as the buffer.